refactor json usage
